% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{cv2Output}
\alias{cv2Output}
\title{Title}
\usage{
cv2Output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{}

\item{width}{}

\item{height}{}
}
\description{
Title
}
\examples{

}
