% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{imread}
\alias{imread}
\title{Overload of OpenCV imread.}
\usage{
imread(filename, flags = -1L)
}
\arguments{
\item{filename}{local filepath or url}

\item{flags}{>0 Return a 3-channel color image. (CV_LOAD_IMAGE_[ANYDEPTH|COLOR|GRAYSCALE])
=0 Return a grayscale image.
<0 Return the loaded image as is (with alpha channel).}
}
\value{
numpy.ndarray that can be displayed with plot()
}
\description{
It allows to open a file or an url
}
\examples{
my_image <- imread("https://www.google.fr/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png")
imshow(my_image)

# you can also do plot(my_image)


}
