% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{imshow}
\alias{imshow}
\title{Overload of OpenCV imshow to make it compatible with RStudio server and Shiny}
\usage{
imshow(winname = "default", mat, render_max_w = 1000,
  render_max_h = 1000, keep_shape = T, scale = 1)
}
\arguments{
\item{winname}{Image identifyer}

\item{mat}{image data structure to show}

\item{render_max_w}{make sure we do not send too much pixels over http}

\item{render_max_h}{make sure we do not send too much pixels over http}

\item{keep_shape}{keep aspect ratio when resizing according to render_max_w|h (default TRUE)}

\item{scale}{Scale factor on client size (does not affect bandwidth)}
}
\value{
r2d3 object
}
\description{
Overload of OpenCV imshow to make it compatible with RStudio server and Shiny
}
\examples{
my_image <- imread("https://upload.wikimedia.org/wikipedia/fr/4/4e/RStudio_Logo.png")
imshow(mat=my_image)

# you can also do plot(my_image)

}
