% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{inputCv2Cam}
\alias{inputCv2Cam}
\title{Capture WebCam}
\usage{
inputCv2Cam(inputId, width = 320, height = 240, fps = 10,
  display = T)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{width}{Width of the captured image}

\item{height}{Height of the captured image}

\item{fps}{Frames per secondes}

\item{show_live}{Shall we show webcam stream}

\item{show_captured}{Shall we show the captured image}
}
\value{
A WebCam capture control that can be added to a UI definition
}
\description{
Capture WebCam
}
\examples{

library(shiny)

ui <- fluidPage(
    inputCv2Cam("video")
)

server <- function(input, output, session) {
    observe({ print(nchar(input$video)) })
}

shinyApp(ui, server)

}
