% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.numpy.ndarray}
\alias{plot.numpy.ndarray}
\title{plot an image from OpenCV}
\usage{
\method{plot}{numpy.ndarray}(mat)
}
\arguments{
\item{mat}{image to plot}
}
\value{
Show image in Viwer pane
}
\description{
plot an image from OpenCV
}
\examples{

my_image <- imread("https://www.google.fr/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png")
plot(my_image)

}
