% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.numpy.ndarray}
\alias{print.numpy.ndarray}
\title{print an image from OpenCV}
\usage{
\method{print}{numpy.ndarray}(mat)
}
\arguments{
\item{mat}{image to plot}
}
\value{
Show image in Viwer pane
}
\description{
print an image from OpenCV
}
\examples{

my_image <- imread("https://www.google.fr/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png")
print(my_image) # or just my_image

}
