% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{imshow}
\alias{imshow}
\title{Overload of OpenCV imshow to make it compatible with RStudio server and Shiny}
\usage{
imshow(img, max_w = 500, max_h = 500, keep_shape = T)
}
\arguments{
\item{img}{image data structure to show}

\item{max_w}{resize with maximum number of pixel width}

\item{max_h}{resize with maximum number of pixel height}

\item{keep_shape}{keep aspect ratio when resizing (default TRUE)}
}
\value{
r2d3 object
}
\description{
Overload of OpenCV imshow to make it compatible with RStudio server and Shiny
}
\examples{
my_image <- imread("https://www.google.fr/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png")
imshow(my_image)

# you can also do plot(my_image)

}
